backend/requirements.txt
fastapi
uvicorn
pydantic
motor
pymongo
python-multipart

python-multipart
python-dotenv
google-generativeai
markdown
PyPDF2
docx2txt
python-docx
docx2pdf
assemblyai
pillow
firebase-admin